+++ adapter/entry-point/src/main/java/org/example/config/BeanConfiguration.java
@@ -15,48 +15,7 @@ public class BeanConfiguration {
     @Bean
     public ClientRepository clientRepository(){
 
-        return new ClientRepository() {
-
-
-            private static final Client client = new Client("C",
-                    "23445322",
-                    "Johnathan",
-                    "Arley",
-                    "Monsalve",
-                    "bello",
-                    "123456789",
-                    "calle san juan",
-                    "med"
-            );
-
-            private Map<String, Client> clients = new HashMap<>(
-                    Map.of(client.idType().concat(client.idNumber()), client)
-            );
-
-
-            @Override
-            public Optional<List<Client>> getClients() {
-                return Optional.of(clients.values().stream().toList());
-            }
-
-            @Override
-            public Optional<Client> getClientByID(String idType, String idNumber) {
-
-                return Optional.ofNullable(clients.get(idType.concat(idNumber)));
-            }
-
-            @Override
-            public Optional<Client> save(Client client) {
-                String key =  client.idType().concat(client.idNumber());
-                clients.put(key,client);
-                return Optional.ofNullable(clients.get(key));
-            }
-
-            @Override
-            public Optional<Client> delete(String idType, String idNumber) {
-                return Optional.ofNullable(clients.remove(idType.concat(idNumber)));
-            }
-        };
+        return new TemporalRepository();
     }
 
     @Bean
